A bit of a refactor for JSON load/store logic for most aux structs, continuing with Cache serialization
